aquila-preprocess/requirements.txt
pandas~=1.3.4
psycopg2~=2.9.2
sqlalchemy~=1.4.27
crossrefapi~=1.5.0
scanpy~=1.8.2
anndata~=0.7.8
aquila~=0.0.0
python-dotenv~=0.19.2
orjson~=3.6.5
rich~=10.15.2